refactor: move tests outside of read.rs
