+++ src/read.rs
@@ -1195,130 +1195,3 @@ pub fn root_dir_common_filter(path: &Path) -> bool {
     true
 }
 
-#[cfg(test)]
-mod tests {
-    use std::io::Cursor;
-
-    /// Only on little endian because we cannot use fs with miri CI
-    #[cfg(all(target_endian = "little", not(miri)))]
-    #[test]
-    fn test_is_symlink() -> std::io::Result<()> {
-        use super::ZipArchive;
-        use tempfile::TempDir;
-
-        let mut reader = ZipArchive::new(Cursor::new(include_bytes!("../tests/data/symlink.zip")))?;
-        assert!(reader.by_index(0)?.is_symlink());
-        let tempdir = TempDir::with_prefix("test_is_symlink")?;
-        reader.extract(&tempdir)?;
-        assert!(tempdir.path().join("bar").is_symlink());
-        Ok(())
-    }
-
-    #[test]
-    #[cfg(feature = "deflate-flate2")]
-    fn test_utf8_extra_field() {
-        use super::ZipArchive;
-
-        let mut reader =
-            ZipArchive::new(Cursor::new(include_bytes!("../tests/data/chinese.zip"))).unwrap();
-        reader.by_name("七个房间.txt").unwrap();
-    }
-
-    #[test]
-    fn test_utf8() {
-        use super::ZipArchive;
-
-        let mut reader =
-            ZipArchive::new(Cursor::new(include_bytes!("../tests/data/linux-7z.zip"))).unwrap();
-        reader.by_name("你好.txt").unwrap();
-    }
-
-    #[test]
-    fn test_utf8_2() {
-        use super::ZipArchive;
-
-        let mut reader = ZipArchive::new(Cursor::new(include_bytes!(
-            "../tests/data/windows-7zip.zip"
-        )))
-        .unwrap();
-        reader.by_name("你好.txt").unwrap();
-    }
-
-    /// Only on little endian because it runs too long with Miri CI
-    #[cfg(all(target_endian = "little", not(miri)))]
-    #[test]
-    fn test_64k_files() -> crate::result::ZipResult<()> {
-        use super::ZipArchive;
-        use crate::CompressionMethod::Stored;
-        use crate::ZipWriter;
-        use crate::types::SimpleFileOptions;
-        use std::io::{Read, Write};
-
-        let mut writer = ZipWriter::new(Cursor::new(Vec::new()));
-        let options = SimpleFileOptions {
-            compression_method: Stored,
-            ..Default::default()
-        };
-        for i in 0..=u16::MAX {
-            let file_name = format!("{i}.txt");
-            writer.start_file(&*file_name, options)?;
-            writer.write_all(i.to_string().as_bytes())?;
-        }
-
-        let mut reader = ZipArchive::new(writer.finish()?)?;
-        for i in 0..=u16::MAX {
-            let expected_name = format!("{i}.txt");
-            let expected_contents = i.to_string();
-            let expected_contents = expected_contents.as_bytes();
-            let mut file = reader.by_name(&expected_name)?;
-            let mut contents = Vec::with_capacity(expected_contents.len());
-            file.read_to_end(&mut contents)?;
-            assert_eq!(contents, expected_contents);
-            drop(file);
-            contents.clear();
-            let mut file = reader.by_index(i as usize)?;
-            file.read_to_end(&mut contents)?;
-            assert_eq!(contents, expected_contents);
-        }
-        Ok(())
-    }
-
-    /// Symlinks being extracted shouldn't be followed out of the destination directory.
-    /// Only on little endian because we cannot use fs with miri CI
-    #[cfg(all(target_endian = "little", not(miri)))]
-    #[test]
-    fn test_cannot_symlink_outside_destination() -> crate::result::ZipResult<()> {
-        use crate::ZipWriter;
-        use crate::types::SimpleFileOptions;
-        use std::fs::create_dir;
-        use tempfile::TempDir;
-
-        let mut writer = ZipWriter::new(Cursor::new(Vec::new()));
-        writer.add_symlink("symlink/", "../dest-sibling/", SimpleFileOptions::default())?;
-        writer.start_file("symlink/dest-file", SimpleFileOptions::default())?;
-        let mut reader = writer.finish_into_readable()?;
-        let dest_parent = TempDir::with_prefix("read__test_cannot_symlink_outside_destination")?;
-        let dest_sibling = dest_parent.path().join("dest-sibling");
-        create_dir(&dest_sibling)?;
-        let dest = dest_parent.path().join("dest");
-        create_dir(&dest)?;
-        assert!(reader.extract(dest).is_err());
-        assert!(!dest_sibling.join("dest-file").exists());
-        Ok(())
-    }
-
-    /// Only on little endian because we cannot use fs with miri CI
-    #[cfg(all(target_endian = "little", not(miri)))]
-    #[test]
-    fn test_can_create_destination() -> crate::result::ZipResult<()> {
-        use super::ZipArchive;
-        use tempfile::TempDir;
-
-        let mut reader =
-            ZipArchive::new(Cursor::new(include_bytes!("../tests/data/mimetype.zip")))?;
-        let dest = TempDir::with_prefix("read__test_can_create_destination")?;
-        reader.extract(&dest)?;
-        assert!(dest.path().join("mimetype").exists());
-        Ok(())
-    }
-}
