+++ tests/extract_symlink.rs
@@ -1,3 +1,21 @@
+/// Only on little endian because we cannot use fs with miri CI
+#[cfg(all(target_endian = "little", not(miri)))]
+#[test]
+fn test_is_symlink() -> std::io::Result<()> {
+    use zip::ZipArchive;
+    use tempfile::TempDir;
+    use std::io::Cursor;
+
+    let mut reader = ZipArchive::new(Cursor::new(include_bytes!("../tests/data/symlink.zip")))?;
+    assert!(reader.by_index(0)?.is_symlink());
+    let tempdir = TempDir::with_prefix("test_is_symlink")?;
+    reader.extract(&tempdir)?;
+    assert!(tempdir.path().join("bar").is_symlink());
+    Ok(())
+}
+
+
+
 #[test]
 #[cfg(all(unix, feature = "deflate-flate2"))]
 fn extract_should_respect_links() {
@@ -18,3 +36,29 @@ fn extract_should_respect_links() {
 
     assert_eq!(target_path, PathBuf::from_str("pandoc").unwrap());
 }
+
+/// Symlinks being extracted shouldn't be followed out of the destination directory.
+/// Only on little endian because we cannot use fs with miri CI
+#[cfg(all(target_endian = "little", not(miri)))]
+#[test]
+fn test_cannot_symlink_outside_destination() -> zip::result::ZipResult<()> {
+    use zip::ZipWriter;
+    use zip::write::SimpleFileOptions;
+    use std::fs::create_dir;
+    use tempfile::TempDir;
+    use std::io::Cursor;
+
+    let mut writer = ZipWriter::new(Cursor::new(Vec::new()));
+    writer.add_symlink("symlink/", "../dest-sibling/", SimpleFileOptions::default())?;
+    writer.start_file("symlink/dest-file", SimpleFileOptions::default())?;
+    let mut reader = writer.finish_into_readable()?;
+    let dest_parent = TempDir::with_prefix("read__test_cannot_symlink_outside_destination")?;
+    let dest_sibling = dest_parent.path().join("dest-sibling");
+    create_dir(&dest_sibling)?;
+    let dest = dest_parent.path().join("dest");
+    create_dir(&dest)?;
+    assert!(reader.extract(dest).is_err());
+    assert!(!dest_sibling.join("dest-file").exists());
+    Ok(())
+}
+
